requirements.txt
pytest==7.2.0
redis==4.4.0
python-telegram-bot==20.0a5
requests==2.28.1
python-dotenv==0.21.0